Distinguish between "last instruction" and "whole list" flags in compiler nodes.
For optimizations, we need to keep the exact flags of the last instruction in an instruction list, while the list itself can have different flags set from earlier instructions.
Even it that's a bit confusing, we only look at the "list" flags downstream to decide whether an expression can be const-evaluated and/or optimized away.